Patched get/updateRecord in PropertyStore and all dynamic stores to allow the higher order bits

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/nioneo/store/DynamicStringStore.java
@@ -59,7 +59,7 @@ public class DynamicStringStore extends AbstractDynamicStore
         super.setHighId( highId );
     }
     
-    public int nextBlockId()
+    public long nextBlockId()
     {
         return super.nextBlockId();
     }
